requirements.txt
pytest-playwright>=0.5.1
pytest-html>=4.1.1
pytest-json-report>=1.5.0
pytest-timeout>=2.3.1
python-dotenv>=1.0.1
pytest-retry>=1.6.3
pytest-xdist>=3.6.1
mypy>=1.14.1
ruff>=0.9.6